Create coefficients module
